Attempt at finding/removing outliers
Did a little reorganization/consolidation, but this is mostly me trying to work on finding and removing outliers.  I started with "pledged" but ran into some issues and am not entirely sure what's happening (or not)?  Goingto take a break and can try again later, but uploading this in case one of you might be able to figure out what the issues are and get it to work before proceeding onto the other two columns.