Update Kickstarter_Combined_CampaignsByStatus_TimeSeries.R
Fixed the outlier work for pledged with Owen's help.

Copied and modified that code for backers and goals.

Still running into some issues with pulling out the low/high outliers for goals and for backers -- see comments for where and feel free to fix if you can see what needs to be done